# CU-39cmvru Faster hashing