--- a/medcat/cat.py
+++ b/medcat/cat.py
@@ -194,13 +194,13 @@ class CAT(object):
         else:
             return json.dumps(card, indent=2, sort_keys=False)
 
-    def _versioning(self):
+    def _versioning(self, force_rehash: bool = False):
         # Check version info and do not allow without it
         if self.config.version.description == 'No description':
             logger.warning("Please consider populating the version information [description, performance, location, ontology] in cat.config.version")
 
         # Fill the stuff automatically that is needed for versioning
-        m = self.get_hash()
+        m = self.get_hash(force_recalc=force_rehash)
         version = self.config.version
         if version.id is None or m != version.id:
             if version.id is not None:
@@ -212,7 +212,7 @@ class CAT(object):
             version.medcat_version = __version__
             logger.warning("Please consider updating [description, performance, location, ontology] in cat.config.version")
 
-    def create_model_pack(self, save_dir_path: str, model_pack_name: str = DEFAULT_MODEL_PACK_NAME) -> str:
+    def create_model_pack(self, save_dir_path: str, model_pack_name: str = DEFAULT_MODEL_PACK_NAME, force_rehash: bool = False) -> str:
         """Will crete a .zip file containing all the models in the current running instance
         of MedCAT. This is not the most efficient way, for sure, but good enough for now.
 
@@ -221,6 +221,8 @@ class CAT(object):
                 An id will be appended to this name
             model_pack_name (str, optional):
                 The model pack name. Defaults to DEFAULT_MODEL_PACK_NAME.
+            force_rehash (bool, optional):
+                Force recalculation of hash. Defaults to `False`.
 
         Returns:
             str:
@@ -229,7 +231,7 @@ class CAT(object):
         # Spacy model always should be just the name, but during loading it can be reset to path
         self.config.general.spacy_model = os.path.basename(self.config.general.spacy_model)
         # Versioning
-        self._versioning()
+        self._versioning(force_rehash)
         model_pack_name += "_{}".format(self.config.version.id)
 
         logger.warning("This will save all models into a zip file, can take some time and require quite a bit of disk space.")