--- a/medcat/cdb.py
+++ b/medcat/cdb.py
@@ -702,7 +702,7 @@ or download the compatible model."""
 
     def get_hash(self, force_recalc: bool = False):
         if not force_recalc and self._hash and not self.is_dirty:
-            logger.info("Resuing old hash of CDB since the CDB has not changed: %s", self._hash)
+            logger.info("Reusing old hash of CDB since the CDB has not changed: %s", self._hash)
             return self._hash
         self.is_dirty = False
         return self.calculate_hash()