Hashing of model packs has been quite slow. This is, in large part, due to the large amount of data being hashed within the CDB. However, in practice, the CDB may not always change. Thus, rehashing it may be a waste of resources (i.e time).

What this PR does is cache the hash for a CDB in memory and track whether or not the CDB has changed (i.e is _dirty_). If the CDB is not dirty, its previous hash is used (if present / known); only if the CDB is dirty (or the hash was not previously known), is the hash recalculated.

The hash is saved on disk, if known. The user can also specify to calculate the hash when saving the CDB if one isn't known. The hash may not be known if the CDB is saved outside a model pack and no hash was read from disk. When saved within a model pack, the hash is known since it is used as part of the model's ID/hash and has thus been calculated.

The PR allows for the user to force the recalculation of the CDB hash as well. Both when saving the model pack as well as when simply getting the hash of the CDB. This is in case the CDB internals (i.e the various dicts) are manipulated outside the API that is provided (for whatever reason; this should not be an issue in normal operation).

PS:
The default behaviour now is to use the cached CDB hash (if present). That is to say, the default behaviour has changed.

PPS:
The PR does make calculating the hash significantly faster. I did some preliminary testing on Anthony's 1.4 model as well as the UMLS model and this is what I found on my M1 MacBook Pro:
Anthony's 1.4 SNOMED model pack:
- CDB hashed: 140 (±2)s
- Using cached CDB hash: <0.001s (i.e <1ms)

UMLS model pack:
- CDB hashed: 690 (±70)s
- Using cached CDB hash: ~12s

So there seems to be significant performance boost.
On the UMLS side, most of the 12s is spent hashing the config (pretty much all of it, I checked). More specifically, on hashing the `config.linking.filters` (around 8s); it contains all the CUIs within the model.